move PrimitiveVectorElement
