+++ lib/segment/src/vector_storage/dense/mod.rs
@@ -3,3 +3,4 @@ mod dynamic_mmap_flags;
 pub mod memmap_dense_vector_storage;
 pub mod mmap_dense_vectors;
 pub mod simple_dense_vector_storage;
+mod primitive;
